Improve the accuracy of the sensor, co2 adds force recalibration and background calibration

+++ main/i2c_task.h
@@ -44,7 +44,7 @@ typedef struct{
 	uint16_t voc_value;
 	uint16_t light_value;
 	uint8_t occ;
-	uint32_t sound;
+	uint16_t sound;
 	uint16_t ambient;
 	uint16_t object;
 	float original_temperature;
